Refactor feat: Improve LLM model selection UI
The changes in this commit focus on improving the user interface for selecting the
LLM model in the settings. The key changes are:

- Refactor the conditional rendering logic to make it more readable and
  maintainable.
- Ensure the "No models available" message is displayed consistently when there
  are no available models.
- Improve the formatting and layout of the LLM model selection UI.
