+++ apps/desktop/src/components/settings/components/ai/index.ts
@@ -1,0 +1,3 @@
+export * from "./llm-view";
+export * from "./stt-view";
+export * from "./wer-modal";
